Fix CI

+++ esp-idf-svc/src/bt.rs
@@ -686,7 +686,7 @@ where
             version_num: unsafe { crate::sys::esp_ble_get_chip_rev_version() },
             #[cfg(esp32c6)]
             #[allow(clippy::unnecessary_cast)]
-            version_num: unsafe { crate::sys::efuse_hal_chip_revision() as u8 },
+            version_num: unsafe { crate::sys::efuse_hal_chip_revision() as u32 },
             #[cfg(not(esp32c2))]
             cpu_freq_mhz: crate::sys::CONFIG_ESP_DEFAULT_CPU_FREQ_MHZ as _,
             ignore_wl_for_direct_adv: 0,
